MODIFY && ADD -- 修改 BlogService 的 getBlogById 参数类型为 long，BlogController 中新增了返回特定博客详细信息的方法

+++ src/main/java/com/jessysnow/boot/service/BlogService.java
@@ -22,5 +22,5 @@ public interface BlogService {
      * @param id 博客 id
      * @return Blog 对象
      */
-    Blog getBlogById(int id);
+    Blog getBlogById(long id);
 }
